models/train_model.py
# Import necessary libraries
import pandas as pd  # Pandas library for data manipulation
import numpy as np  # NumPy library for numerical operations
from sklearn.model_selection import train_test_split  # Scikit-Learn library for data splitting
from sklearn.ensemble import RandomForestClassifier  # Random Forest Classifier
from sklearn.metrics import accuracy_score  # Scikit-Learn library for model evaluation

# Load historical data
data = pd.read_csv("../data/historical_data.csv")

# Prepare data for training
# Your data preprocessing and feature engineering code here
# (This is where you would typically prepare your dataset, clean and transform data, and engineer features)

# Split data into training and testing sets
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)
# - X: Features (independent variables)
# - y: Target variable (dependent variable)
# - test_size: The proportion of the dataset to include in the test split (20% in this case)
# - random_state: Seed for random number generator to ensure reproducibility

# Train a machine learning model (e.g., Random Forest)
model = RandomForestClassifier(n_estimators=100)
# - n_estimators: Number of trees in the random forest (100 in this case)

model.fit(X_train, y_train)
# - Fit (train) the model using the training data

# Make predictions
y_pred = model.predict(X_test)
# - Predict the target variable using the test set

# Evaluate model accuracy
accuracy = accuracy_score(y_test, y_pred)
# - Calculate the accuracy of the model's predictions compared to the actual values in the test set
# - accuracy_score() computes the accuracy of the classification model

print(f"Model Accuracy: {accuracy}")
# - Print the accuracy of the trained model


""" 
Explanation:

Import necessary libraries: Import the required libraries, including pandas, numpy, train_test_split from Scikit-Learn for data splitting, RandomForestClassifier from Scikit-Learn for a Random Forest model, and accuracy_score from Scikit-Learn for model evaluation.

Load historical data: Load historical data from a CSV file located at "../data/historical_data.csv" into a Pandas DataFrame called data. This assumes that your data file is in the specified location.

Data preparation: This is a placeholder comment where you would typically include data preprocessing and feature engineering code. In this section, you would clean, transform, and engineer features in your dataset to prepare it for training.

Split data into training and testing sets: Use the train_test_split function to split your data into training (X_train and y_train) and testing (X_test and y_test) sets. This is crucial for evaluating your model's performance.

Train a machine learning model: Create a Random Forest Classifier model with 100 trees and train it using the training data (X_train and y_train).

Make predictions: Use the trained model to make predictions on the test data (X_test) and store the predicted values in y_pred.

Evaluate model accuracy: Calculate the accuracy of the model's predictions by comparing them to the actual target values (y_test). The accuracy_score function computes the accuracy, which represents the fraction of correctly classified instances.

Print model accuracy: Finally, print the accuracy of the trained model on the test set to assess its performance.
"""