strategies/trading_strategy.py
# Import necessary libraries
import os  # Operating System module for handling environment variables
import alpaca_trade_api as tradeapi  # Alpaca Trade API library for trading
from dotenv import load_dotenv  # Load environment variables from a .env file

# Load environment variables from a .env file (assuming you have one)
load_dotenv()

# Retrieve the Alpaca API key and secret key from environment variables
ALPACA_API_KEY = os.getenv("ALPACA_API_KEY")
ALPACA_SECRET_KEY = os.getenv("ALPACA_SECRET_KEY")

# Initialize the Alpaca REST API client with the provided API keys
api = tradeapi.REST(ALPACA_API_KEY, ALPACA_SECRET_KEY, base_url="https://paper-api.alpaca.markets")
# - The 'base_url' is set to the Alpaca paper trading environment

# Import your machine learning model and preprocessing functions
from your_model_file import predict_price_movement, preprocess_data
# - Assuming you have a separate Python file ('your_model_file.py') containing your machine learning model and data preprocessing functions

# Define the stock symbol you want to trade
symbol = "AAPL"

def execute_trades():
    # Fetch historical data for the stock using the Alpaca API
    historical_data = api.get_barset(symbol, 'day', limit=20).df[symbol]
    # - This fetches the last 20 days of daily trading data for the specified symbol
    
    # Prepare data for prediction using your preprocessing function
    X = preprocess_data(historical_data)
    # - Call your preprocessing function to prepare the data for making predictions
    
    # Make predictions using your machine learning model
    prediction = predict_price_movement(X)
    # - Call your machine learning model to predict the price movement
    
    # Determine the trading action based on the prediction
    if prediction == 1:  # Positive prediction, buy
        api.submit_order(
            symbol=symbol,
            qty=1,  # You can adjust the quantity based on your strategy
            side='buy',
            type='market',
            time_in_force='gtc'
        )
        print(f"Bought 1 share of {symbol}")
    elif prediction == -1:  # Negative prediction, sell
        api.submit_order(
            symbol=symbol,
            qty=1,  # You can adjust the quantity based on your strategy
            side='sell',
            type='market',
            time_in_force='gtc'
        )
        print(f"Sold 1 share of {symbol}")
    else:
        print(f"No action taken for {symbol}")
        # - If the prediction is neither 1 (buy) nor -1 (sell), no action is taken

if __name__ == "__main__":
    execute_trades()
    # - This block of code will only run if the script is executed directly, not when it's imported as a module


""" 
Explanation:

Import necessary libraries: Import the required libraries, including os, alpaca_trade_api for Alpaca trading, and load_dotenv to load environment variables from a .env file.

Load environment variables: Load the Alpaca API key and secret key from environment variables for secure authentication with the Alpaca API.

Initialize the Alpaca REST API client: Create an Alpaca REST API client instance using the provided API keys, and set the base_url to the Alpaca paper trading environment.

Import your machine learning model and preprocessing functions: Import your machine learning model and any necessary data preprocessing functions from a separate Python file (assumed to be named 'your_model_file.py').

Define the stock symbol: Set the variable symbol to the stock symbol you want to trade, in this case, "AAPL" for Apple Inc.

Define the execute_trades function: This function fetches historical trading data for the specified symbol, preprocesses the data, makes predictions using your machine learning model, and executes trading orders based on the predictions.

Fetch historical data: Use the Alpaca API to fetch the last 20 days of daily trading data for the specified symbol ("AAPL") and store it in the historical_data DataFrame.

Prepare data for prediction: Call your preprocess_data function to prepare the historical data for making predictions. This step may involve data cleaning, feature engineering, and scaling, depending on your model's requirements.

Make predictions: Use your predict_price_movement function to predict the price movement based on the preprocessed data. The model's output (prediction) is expected to be 1 for a buy signal, -1 for a sell signal, or any other value for no action.

Determine the trading action: Based on the prediction, execute a buy or sell order using the Alpaca API. The api.submit_order function is used to place market orders with a quantity of 1 share, but you can adjust the quantity and order parameters based on your trading strategy.

Print trade actions: Print the actions taken by the script, such as buying or selling shares, or indicating no action, for the specified symbol.

Conditional execution: The if __name__ == "__main__": block ensures that the execute_trades function is executed only when the script is run directly, not when it's imported as a module in another script.
"""