requirements.txt
pandas
numpy
scikit-learn
tensorflow 
pytorch
alpha_vantage
alpaca-trade-api
python-dotenv
